+++ NEWS.md
@@ -1,3 +1,5 @@
+# fs (development version)
+
 # fs 1.4.2
 
 * `file_info(..., follow = TRUE)`, `is_dir()`, and `is_file()`